+++ src/datasources/results/ResultsDataSourceBase.ts
@@ -1,8 +1,7 @@
 import { DataSourceBase } from "core/DataSourceBase";
 import { DataQueryRequest, DataFrameDTO, TestDataSourceResponse } from "@grafana/data";
-import { ResultsQuery } from "./types/types";
+import { ProductProperties, QueryProductResponse, ResultsQuery } from "./types/types";
 import { QueryBuilderOption, Workspace } from "core/types";
-import { ResultsPropertiesOptions } from "./types/QueryResults.types";
 import { getVariableOptions } from "core/utils";
 import { ExpressionTransformFunction } from "core/query-builder.utils";
 import { QueryBuilderOperations } from "core/query-builder.constants";
@@ -10,6 +9,7 @@ import { QueryBuilderOperations } from "core/query-builder.constants";
 export abstract class ResultsDataSourceBase extends DataSourceBase<ResultsQuery> {
   baseUrl = this.instanceSettings.url + '/nitestmonitor';
   queryResultsValuesUrl = this.baseUrl + '/v2/query-result-values';
+  queryProductsUrl = this.baseUrl + '/v2/query-products';
 
   private timeRange: { [key: string]: string } = {
     Started: 'startedAt',
@@ -19,7 +19,7 @@ export abstract class ResultsDataSourceBase extends DataSourceBase<ResultsQuery>
   private fromDateString = '${__from:date}';
   private toDateString = '${__to:date}';
   private static _workspacesCache: Promise<Map<string, Workspace>> | null = null;
-  private static _partNumbersCache: Promise<string[]> | null = null;
+  private static _productCache: Promise<QueryProductResponse> | null = null;
 
   readonly globalVariableOptions = (): QueryBuilderOption[] => getVariableOptions(this);
 
@@ -43,8 +43,8 @@ export abstract class ResultsDataSourceBase extends DataSourceBase<ResultsQuery>
     return this.loadWorkspaces();
   }
 
-  get partNumbersCache(): Promise<string[]> {
-    return this.getPartNumbers();
+  get productCache(): Promise<QueryProductResponse> {
+    return this.loadProducts();
   }
 
   async loadWorkspaces(): Promise<Map<string, Workspace>> {
@@ -68,20 +68,6 @@ export abstract class ResultsDataSourceBase extends DataSourceBase<ResultsQuery>
     return ResultsDataSourceBase._workspacesCache;
   }
 
-  async getPartNumbers(): Promise<string[]> {
-    if (ResultsDataSourceBase._partNumbersCache) {
-      return ResultsDataSourceBase._partNumbersCache;
-    }
-
-    ResultsDataSourceBase._partNumbersCache = this.queryResultsValues(ResultsPropertiesOptions.PART_NUMBER, undefined)
-    .catch(error => {
-      console.error('Error in loading part numbers:', error);
-      return [];
-    });
-
-    return ResultsDataSourceBase._partNumbersCache;
-  }
-
   async queryResultsValues(fieldName: string, filter?: string): Promise<string[]> {
     try {
       return await this.post<string[]>(this.queryResultsValuesUrl, {
@@ -92,6 +78,31 @@ export abstract class ResultsDataSourceBase extends DataSourceBase<ResultsQuery>
       throw new Error(`An error occurred while querying result values: ${error}`);
     }
   }
+
+  async queryProducts(
+    projection?: ProductProperties[],
+  ): Promise<QueryProductResponse> {
+    try {
+      const response = await this.post<QueryProductResponse>(this.queryProductsUrl, {
+        projection,
+      });
+      return response;
+    } catch (error) {
+      throw new Error(`An error occurred while querying products: ${error}`);
+    }
+  }
+
+  async loadProducts(): Promise<QueryProductResponse> {
+    if (ResultsDataSourceBase._productCache) {
+      return ResultsDataSourceBase._productCache;
+    }
+    ResultsDataSourceBase._productCache = this.queryProducts([ProductProperties.name, ProductProperties.partNumber])
+      .catch(error => {
+        console.error('Error in loading products:', error);
+        return { products: [] };
+      });
+    return ResultsDataSourceBase._productCache;
+  };
 
   protected multipleValuesQuery(field: string): ExpressionTransformFunction {
     return (value: string, operation: string, _options?: any) => {
@@ -119,6 +130,21 @@ export abstract class ResultsDataSourceBase extends DataSourceBase<ResultsQuery>
   protected buildQueryFilter(filterA?: string, filterB?: string): string | undefined {
     const filters = [filterA, filterB].filter(Boolean);
     return filters.length > 0 ? filters.join(' && ') : undefined;
+  };
+
+  /**
+   * Builds a query string for the given field using multiple values,
+   * joining each condition with the logical OR operator
+   *
+   * @param fieldName - The field name to filter on
+   * @param values - Array of values to include in the OR condition.
+   * @returns The constructed query string, or an empty string if no values are provided.
+   */
+  protected buildQueryWithOrOperator = (fieldName: string, values: string[]): string => {
+    if (values.length === 0){
+      return '';
+    } 
+    return values.map(item => `${fieldName} = "${item}"`).join(' || ');
   };
 
   private isMultiSelectValue(value: string): boolean {