+++ src/datasources/results/components/editors/query-steps/QueryStepsEditor.test.tsx
@@ -28,6 +28,18 @@ jest.mock('../../query-builders/steps-querybuilder-wrapper/StepsQueryBuilderWrap
   }),
 }));
 
+const mockProducts = {
+  products: [
+    {partNumber: 'PartNumber1', name: 'ProductName1'},
+    {partNumber: 'PartNumber2', name: 'ProductName2'}
+  ]
+}
+
+const mockGlobalVars = [
+  { label: '$var1', value: '$var1' },
+  { label: '$var2', value: '$var2' }
+];
+
 describe('QueryStepsEditor', () => {
   const defaultQuery: QuerySteps = {
     refId: 'A',
@@ -40,6 +52,7 @@ describe('QueryStepsEditor', () => {
     useTimeRangeFor: 'Updated',
     recordCount: 1000,
     showMeasurements: false,
+    partNumberQuery: ['PartNumber1'],
     resultsQuery: 'partNumber = "PN1"',
     stepsQuery: 'stepName = "Step1"',
   };
@@ -49,9 +62,9 @@ describe('QueryStepsEditor', () => {
   const mockDatasource = {
     loadWorkspaces: jest.fn(),
     getPartNumbers: jest.fn(),
+    productCache: Promise.resolve(mockProducts),
     workspacesCache: new Map(),
-    partNumbersCache: [],
-    globalVariableOptions: jest.fn(() => []),
+    globalVariableOptions: jest.fn(() => mockGlobalVars),
     disableStepsQueryBuilder: false
   } as unknown as QueryStepsDataSource;
 
@@ -62,16 +75,20 @@ describe('QueryStepsEditor', () => {
   let dataOutput: HTMLElement;
   let totalCountOutput: HTMLElement;
   let showMeasurements: HTMLElement;
+  let productName: HTMLElement;
 
-  beforeEach(() => {
-    render(<QueryStepsEditor query={defaultQuery} handleQueryChange={mockHandleQueryChange} datasource={mockDatasource}/>);
+  beforeEach(async () => {
+    await act(async () => {
+      render(<QueryStepsEditor query={defaultQuery} handleQueryChange={mockHandleQueryChange} datasource={mockDatasource}/>);
+    });
     properties = screen.getAllByRole('combobox')[0];
-    orderBy = screen.getAllByRole('combobox')[2];
+    orderBy = screen.getAllByRole('combobox')[3];
     descending = screen.getAllByRole('checkbox')[2];
     dataOutput = screen.getByRole('radio', { name: 'Data' });
     totalCountOutput = screen.getByRole('radio', { name: 'Total Count' });
     recordCount = screen.getByDisplayValue(1000);
     showMeasurements = screen.getAllByRole('checkbox')[0];
+    productName = screen.getAllByRole('combobox')[2];
   });
 
   describe('Data outputType', () => {
@@ -100,6 +117,8 @@ describe('QueryStepsEditor', () => {
       expect(screen.getAllByText('Updated').length).toBe(1);
       expect(showMeasurements).toBeInTheDocument();
       expect(showMeasurements).not.toBeChecked();
+      expect(productName).toBeInTheDocument();
+      expect(screen.getAllByText('ProductName1 (PartNumber1)').length).toBe(1);
     });
 
     test('should display placeholders for properties and orderBy when default values are not provided', async () => {
@@ -134,6 +153,20 @@ describe('QueryStepsEditor', () => {
       await userEvent.click(descending);
       await waitFor(() => {
         expect(mockHandleQueryChange).toHaveBeenCalledWith(expect.objectContaining({ descending: false }));
+      });
+    });
+
+    test('should update part number query when user selects a variable in product name dropdown', async () => {
+      await select(productName, '$var1', { container: document.body });
+      await waitFor(() => {
+        expect(mockHandleQueryChange).toHaveBeenCalledWith(expect.objectContaining({ partNumberQuery: ["PartNumber1", "$var1"] }));
+      });
+    });
+
+    test('should update part number query when user selects a product in product name dropdown', async () => {
+      await select(productName, 'ProductName2 (PartNumber2)', { container: document.body });
+      await waitFor(() => {
+        expect(mockHandleQueryChange).toHaveBeenCalledWith(expect.objectContaining({ partNumberQuery: ["PartNumber1", "PartNumber2"] }));
       });
     });
 
@@ -181,6 +214,7 @@ describe('QueryStepsEditor', () => {
                 refId: 'A',
                 queryType: QueryType.Steps,
                 outputType: outputType,
+                partNumberQuery: ['partNumber1'],
                 resultsQuery: 'PartNumber = "partNumber1"'
               }}
               handleQueryChange={mockHandleQueryChange}
@@ -227,21 +261,48 @@ describe('QueryStepsEditor', () => {
       );
     });
 
-    test('should handle empty results query and disable steps query builder', async () => {
-      const resultsQueryInput = screen.getByTestId('results-query');
-      const stepsQueryInput = screen.getByTestId('steps-query');
-  
-      fireEvent.change(resultsQueryInput, { target: { value: 'initial-results-query' } });
-      fireEvent.change(stepsQueryInput, { target: { value: 'initial-steps-query' } });
-      fireEvent.change(resultsQueryInput, { target: { value: '' } });
-      
-      await waitFor(() => {
-        expect(mockHandleQueryChange).toHaveBeenCalledWith(
-          expect.objectContaining({ resultsQuery: '' }),
-          false
+    test('should disable steps query builder when partnumber is empty', async () => {
+      cleanup();
+        await act(async () => {
+          render(
+            <QueryStepsEditor
+              query={{
+                refId: 'A',
+                queryType: QueryType.Steps,
+                outputType: OutputType.Data,
+                partNumberQuery: [],
+                stepsQuery: 'stepName = "Step1"',
+              }}
+              handleQueryChange={mockHandleQueryChange}
+              datasource={mockDatasource}
+            />
+          );
+        });
+        const stepsQueryInput = screen.getByTestId('steps-query');
+
+        expect(stepsQueryInput).toBeDisabled();
+    });
+
+    test('should not disable steps query builder when partnumber is not empty', async () => {
+      cleanup();
+      await act(async () => {
+        render(
+          <QueryStepsEditor
+            query={{
+              refId: 'A',
+              queryType: QueryType.Steps,
+              outputType: OutputType.Data,
+              partNumberQuery: ['PartNumber1'],
+              stepsQuery: 'stepName = "Step1"',
+            }}
+            handleQueryChange={mockHandleQueryChange}
+            datasource={mockDatasource}
+          />
         );
       });
-      expect(stepsQueryInput).toBeDisabled();
+      const stepsQueryInput = screen.getByTestId('steps-query');
+
+      expect(stepsQueryInput).not.toBeDisabled();
     });
   })