+++ src/datasources/test-plans/components/TestPlansVariableQueryEditor.tsx
@@ -5,12 +5,15 @@ import { AutoSizeInput, InlineField, InlineSwitch, Select, VerticalGroup } from 
 import { validateNumericInput } from 'core/utils';
 import { TestPlansDataSource } from '../TestPlansDataSource';
 import { TestPlansQueryBuilder } from './query-builder/TestPlansQueryBuilder';
+import { recordCountErrorMessages, TAKE_LIMIT } from '../constants/QueryEditor.constants';
 import { Workspace } from 'core/types';
 
 type Props = QueryEditorProps<TestPlansDataSource, TestPlansVariableQuery>;
 
 export function TestPlansVariableQueryEditor({ query, onChange, datasource }: Props) {
   query = datasource.prepareQuery(query);
+  const [recordCountInvalidMessage, setRecordCountInvalidMessage] = useState<string>('');
+
 
   const [workspaces, setWorkspaces] = useState<Workspace[] | null>(null);
 
@@ -39,11 +42,17 @@ export function TestPlansVariableQueryEditor({ query, onChange, datasource }: Pr
 
   const recordCountChange = (event: React.FormEvent<HTMLInputElement>) => {
     const value = parseInt((event.target as HTMLInputElement).value, 10);
-    if (isNaN(value) || value < 0 || value > 10000) {
-      setIsRecordCountValid(false);
-    } else {
-      setIsRecordCountValid(true);
-      handleQueryChange({ ...query, recordCount: value });
+    switch (true) {
+      case isNaN(value) || value < 0:
+        setRecordCountInvalidMessage(recordCountErrorMessages.greaterOrEqualToZero);
+        break;
+      case value > TAKE_LIMIT:
+        setRecordCountInvalidMessage(recordCountErrorMessages.lessOrEqualToTenThousand);
+        break;
+      default:
+        setRecordCountInvalidMessage('');
+        handleQueryChange({ ...query, recordCount: value });
+        break;
     }
   };
 
@@ -53,8 +62,6 @@ export function TestPlansVariableQueryEditor({ query, onChange, datasource }: Pr
       handleQueryChange({ ...query, queryBy });
     }
   };
-
-  const [isRecordCountValid, setIsRecordCountValid] = useState<boolean>(true);
 
   return (
     <VerticalGroup>
@@ -88,8 +95,8 @@ export function TestPlansVariableQueryEditor({ query, onChange, datasource }: Pr
         label="Take"
         labelWidth={25}
         tooltip={tooltips.recordCount}
-        invalid={!isRecordCountValid}
-        error={errors.recordCount}
+        invalid={!!recordCountInvalidMessage}
+        error={recordCountInvalidMessage}
       >
         <AutoSizeInput
           minWidth={26}
@@ -110,8 +117,4 @@ const tooltips = {
   descending: 'This toggle returns the test plans query in descending order.',
   recordCount: 'This field specifies the maximum number of test plans to return.',
   queryBy: 'This optional field specifies the query filters.'
-};
-
-const errors = {
-  recordCount: 'Record count must be less than 10000'
 };